pyproject.toml
[project]
name = "mcp2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp",
    "sqlmodel",
    "alembic",
    "psycopg",
    "python-dotenv"
]